# perf(thrift): new pure go Skip impl, 1x better performance