--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ output
 .vscode
 *.dSYM
 *.profile
+*.txt