--- a/go.mod
+++ b/go.mod
@@ -1,6 +1,6 @@
 module github.com/cloudwego/dynamicgo
 
-go 1.16
+go 1.17
 
 require (
 	github.com/apache/thrift v0.13.0
@@ -16,4 +16,15 @@ require (
 	github.com/klauspost/cpuid/v2 v2.2.4
 	github.com/stretchr/testify v1.8.2
 	google.golang.org/protobuf v1.28.1
+)
+
+require (
+	github.com/cloudwego/iasm v0.2.0 // indirect
+	github.com/golang/protobuf v1.5.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
+	golang.org/x/arch v0.0.0-20210923205945-b76863e36670 // indirect
+	golang.org/x/sys v0.0.0-20220722155257-8c9f86f7a55f // indirect
+	google.golang.org/genproto v0.0.0-20200526211855-cb27e3aa2013 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )