--- a/thrift/binary_skip_other.go
+++ b/thrift/binary_skip_other.go
@@ -1,7 +1,7 @@
-//go:build amd64
+//go:build !amd64
 
 /**
- * Copyright 2023 cloudwego Inc.
+ * Copyright 2024 cloudwego Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,28 +18,7 @@
 
 package thrift
 
-import (
-	"io"
-
-	"github.com/cloudwego/dynamicgo/internal/native"
-	"github.com/cloudwego/dynamicgo/internal/native/types"
-)
-
 // Skip skips over teh value for the given type using native C implementation.
 func (p *BinaryProtocol) SkipNative(fieldType Type, maxDepth int) (err error) {
-	if maxDepth >= types.TB_SKIP_STACK_SIZE {
-		return errExceedDepthLimit
-	}
-	left := len(p.Buf) - p.Read
-	if left <= 0 {
-		return io.EOF
-	}
-	fsm := types.NewTStateMachine()
-	ret := native.TBSkip(fsm, &p.Buf[p.Read], left, uint8(fieldType))
-	if ret < 0 {
-		return
-	}
-	p.Read += int(ret)
-	types.FreeTStateMachine(fsm)
-	return nil
+	return p.SkipGo(fieldType, maxDepth)
 }