```
goos: darwin
goarch: amd64
pkg: github.com/cloudwego/dynamicgo/thrift
cpu: VirtualApple @ 2.50GHz
                      │  ./old.txt  │              ./new.txt              │
                      │   sec/op    │   sec/op     vs base                │
SkipNoCheck/native-10   185.9n ± 1%   172.6n ± 1%   -7.13% (p=0.000 n=10)
SkipNoCheck/go-10       199.2n ± 1%   100.4n ± 1%  -49.62% (p=0.000 n=10)
geomean                 192.4n        131.6n       -31.60%

                      │  ./old.txt   │              ./new.txt               │
                      │     B/s      │     B/s       vs base                │
SkipNoCheck/native-10   333.6Mi ± 1%   359.1Mi ± 1%   +7.65% (p=0.000 n=10)
SkipNoCheck/go-10       311.1Mi ± 1%   617.9Mi ± 1%  +98.58% (p=0.000 n=10)
geomean                 322.2Mi        471.0Mi       +46.21%
```

#### What type of PR is this?
perf
<!--
Add one of the following kinds:

build: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
ci: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
docs: Documentation only changes
feat: A new feature
optimize: A new optimization
fix: A bug fix
perf: A code change that improves performance
refactor: A code change that neither fixes a bug nor adds a feature
style: Changes that do not affect the meaning of the code (white space, formatting, missing semi-colons, etc)
test: Adding missing tests or correcting existing tests
chore: Changes to the build process or auxiliary tools and libraries such as documentation generation
-->

#### Check the PR title.
<!--
The description of the title will be attached in Release Notes, 
so please describe it from user-oriented, what this PR does / why we need it.
Please check your PR title with the below requirements:
-->
- [ ] This PR title match the format: \<type\>(optional scope): \<description\>
- [ ] The description of this PR title is user-oriented and clear enough for others to understand.
- [ ] Attach the PR updating the user documentation if the current PR requires user awareness at the usage level. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)


#### (Optional) Translate the PR title into Chinese.
对 纯Go版本的 thrift skip 进行性能优化， 一倍性能提升。

#### (Optional) More detailed description for this PR(en: English/zh: Chinese).
<!--
Provide more detailed info for review(e.g., it's recommended to provide perf data if this is a perf type PR).
-->
en:
zh(optional): 


#### (Optional) Which issue(s) this PR fixes:
<!--
Automatically closes linked issue when PR is merged.
Eg: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

#### (optional) The PR that updates user documentation:
<!--
If the current PR requires user awareness at the usage level, please submit a PR to update user docs. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)
-->